// scrapingData/interfaces/Section.ts
export interface CourseInstance {

    department: string;

    courseNum: number;

    section: string;

    title: string;

    crn: number;

    campus: string;

    creditHours: number;

    enrollment: string;

    delivery: string;

    daysOfWeek: string[];

    startTime: number;

    endTime: number;

    buildingName: string;

    roomNumber: number;

    startDate: string;

    endDate: string;
}